# products/src/main/resources/openapi/products-api.yaml
openapi: 3.0.3
info:
  title: Products API
  description: API for product management with CQRS pattern
  version: 1.0.0
servers:
  - url: http://localhost:8080/products
tags:
  - name: Products Query
    description: Read operations (CQRS Query side)
  - name: Products Command
    description: Write operations (CQRS Command side)
paths:
  /:
    get:
      tags:
        - Products Query
      summary: Get service info
      description: Returns basic service information
      operationId: getServiceInfo
      responses:
        '200':
          description: Service information
          content:
            text/plain:
              schema:
                type: string
  /health:
    get:
      tags:
        - Products Query
      summary: Health check
      description: Returns service health status
      operationId: getHealth
      responses:
        '200':
          description: Health status
          content:
            text/plain:
              schema:
                type: string
  /products:
    get:
      tags:
        - Products Query
      summary: Get all products
      description: Returns list of all products
      operationId: getAllProducts
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products Command
      summary: Create product
      description: Creates a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate product name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      tags:
        - Products Query
      summary: Get product by ID
      description: Returns a single product by its ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Products Command
      summary: Update product
      description: Updates an existing product (partial update supported)
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Products Command
      summary: Delete product
      description: Deletes a product by its ID
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cache/evict:
    post:
      tags:
        - Products Command
      summary: Evict all caches
      description: Clears Redis cache for products (forces fresh read from database)
      operationId: evictCaches
      responses:
        '200':
          description: Caches evicted
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    Product:
      type: object
      required:
        - name
        - price
      properties:
        id:
          type: string
          readOnly: true
          description: MongoDB ObjectId
        name:
          type: string
          minLength: 1
          maxLength: 100
        price:
          type: number
          format: double
          minimum: 0.01
          exclusiveMinimum: false
        description:
          type: string
          maxLength: 500
        category:
          type: string
          maxLength: 50
    ProductUpdate:
      type: object
      description: Partial update - all fields optional
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        price:
          type: number
          format: double
          minimum: 0.01
        description:
          type: string
          maxLength: 500
        category:
          type: string
          maxLength: 50
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time